Move the rabit poll helper.
Cleanup all references to the rabit directory.
