+++ tests/ci_build/tidy.py
@@ -192,7 +192,6 @@ class ClangTidy(object):
         def should_lint(path):
             if not self.cpp_lint and path.endswith('.cc'):
                 return False
-            isxgb = path.find('rabit') == -1
             isxgb = isxgb and path.find('dmlc-core') == -1
             isxgb = isxgb and (not path.startswith(self.cdb_path))
             if isxgb:
